+++ MameBake3DLib/CPP/polymesh4.cpp
@@ -528,7 +528,7 @@ int CPolyMesh4::SetPm3InfNoSkin( ID3D11Device* pdev, CMQOObject* srcobj, int clu
 			int vi[3] = {0, 2, 1};
 			int vcnt;
 			for( vcnt = 0; vcnt < 3; vcnt++ ){
-				PM3INF* curinf = m_pm3inf + (int)((INT64)setno * 3 + vcnt);
+				PM3INF* curinf = m_pm3inf + ((size_t)setno * 3 + vcnt);
 				int vno = (m_triface + setno)->GetIndex( vi[vcnt] );
 				_ASSERT( (vno >= 0) && (vno < m_orgpointnum) );
 				curinf->boneindex[0] = clusterno;
@@ -569,7 +569,7 @@ int CPolyMesh4::SetPm3Inf( CMQOObject* srcobj )
 			int vi[3] = {0, 2, 1};
 			int vcnt;
 			for( vcnt = 0; vcnt < 3; vcnt++ ){
-				PM3INF* curinf = m_pm3inf + (int)((INT64)setno * 3 + vcnt);
+				PM3INF* curinf = m_pm3inf + ((size_t)setno * 3 + vcnt);
 				int vno = (m_triface + setno)->GetIndex( vi[vcnt] );
 				_ASSERT( (vno >= 0) && (vno < m_orgpointnum) );
 				INFDATA* curinfdata = (m_infbone + vno)->GetInfData(srcobj);
